+++ test/client_tests/client_test.dart
@@ -19,8 +19,8 @@ import 'package:grpc/grpc.dart';
 import 'package:http2/transport.dart';
 import 'package:test/test.dart';
 
-import 'src/client_utils.dart';
-import 'src/utils.dart';
+import '../src/client_utils.dart';
+import '../src/utils.dart';
 
 void main() {
   const dummyValue = 0;